requirements.txt
flask
openai
packaging
google-cloud-secret-manager
gspread
google-auth
boto3
